requirements.txt
argcomplete==1.12.3
astor==0.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
cachetools==4.2.4
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
colorlog==6.4.1
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
distlib==0.3.2
dl-translate==0.2.6
easydict==1.9
et-xmlfile==1.1.0
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.20
grpcio==1.47.0
grpcio-tools==1.47.0
gunicorn==20.1.0
h5py==3.4.0
huggingface-hub==0.8.1
identify==2.2.14
idna==3.2
imageio==2.13.3
imgaug==0.4.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
lmdb==1.2.1
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
multiprocess==0.70.12.2
networkx==2.6.3
nltk==3.7
nodeenv==1.6.0
numpy==1.21.2
opencv-contrib-python==4.4.0.46
opencv-python==4.5.3.56
openpyxl==3.0.9
packaging==21.0
paddle2onnx==0.7
paddlehub==2.1.0
paddlenlp==2.0.8
paddlepaddle==2.1.2
pandas==1.3.2
Pillow==8.3.2
pipx==0.16.4
platformdirs==2.3.0
pre-commit==2.15.0
premailer==3.10.0
protobuf==3.17.3
pyclipper==1.3.0
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
pyparsing==3.0.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==22.2.1
rarfile==4.0
regex==2022.7.25
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.7.1
sentencepiece==0.1.97
seqeval==1.2.2
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.16.0
smmap==4.0.0
threadpoolctl==2.2.0
tifffile==2021.11.2
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
tqdm==4.62.2
transformers==4.21.1
typing-extensions==3.10.0.2
urllib3==1.26.6
userpath==1.7.0
virtualenv==20.7.2
visualdl==2.2.1
Werkzeug==2.0.1
